+++ src/Handler/PublishCommandHandler.php
@@ -192,7 +192,7 @@ class PublishCommandHandler
     public function handleInstall(Args $args, IO $io)
     {
         if ($args->isArgumentSet('server')) {
-            $expr = Expr::same($args->getArgument('server'), AssetMapping::SERVER_NAME);
+            $expr = Expr::method('getServerName', Expr::same($args->getArgument('server')));
             $mappings = $this->assetManager->findAssetMappings($expr);
         } else {
             $mappings = $this->assetManager->getAssetMappings();
@@ -268,7 +268,7 @@ class PublishCommandHandler
      */
     private function getMappingByUuidPrefix($uuidPrefix)
     {
-        $expr = Expr::startsWith($uuidPrefix, AssetMapping::UUID);
+        $expr = Expr::method('getUuid', Expr::startsWith($uuidPrefix));
 
         $mappings = $this->assetManager->findAssetMappings($expr);